examples/python_fullstack/llama_deploy_app/workflows/requirements.txt
llama-deploy>=0.1.0b3
llama-index-postprocessor-rankgpt-rerank>=0.2.0
llama-index-vector-stores-qdrant>=0.3.0
llama-index-llms-openai>=0.2.2
llama-index-embeddings-openai>=0.2.4
llama-index-readers-file>=0.2.0
